globetrotter-backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-CORS
python-dotenv
ollama
pytest
pytest-flask
Flask-WTF
